[BEAM-3131] Proper icons on theme manager action toolbar
