--- client/Packages/com.beamable/Editor/UI/Buss/ThemeManager/BussThemeManagerActionBarVisualElement/BussThemeManagerActionBarVisualElement.light.uss
@@ -1,23 +1,0 @@
-﻿#root {
-    background-image: resource("Packages/com.beamable/Editor/UI/Content/Icons/actionBarlight.png");
-}
-
-.boundedButton {
-    border-color: #b4b4b4;
-}
-
-#addStyleIcon{
-    background-image: resource("Packages/com.beamable/Editor/UI/Toolbox/Icons/Content_Manager_dark.png");
-}
-
-#duplicateStylesheetIcon{
-    background-image: resource("Packages/com.beamable/Editor/UI/Toolbox/Icons/Content_Manager_dark.png");
-}
-
-#refreshIcon {
-    background-image: resource("Packages/com.beamable/Editor/UI/Content/Icons/Info_dark.png")
-}
-
-#docIcon{
-    background-image: resource("Packages/com.beamable/Editor/UI/Toolbox/Icons/Info_dark.png");
-}
